Allow ignoring of checkstyle
In some cases the checkstyle needs to be worked around where naming
conventions may cause breaking API changes or the checkstyle deviates
from the code formatter where the checkstyle cannot be updated to
reflect the code style.

This change allows you to add a comment in the bit of code the
checkstyle is failing to ignore certain rules e.g.
```java
  // CHECKSTYLE IGNORE Indentation FOR NEXT 1 LINES
```